+++ Sources/SentryObjCCompat/SentryObjCMechanism.swift
@@ -29,7 +29,9 @@ import Foundation
 
     @objc public var data: [String: Any]? {
         get { wrapped.data }
-        set { wrapped.data = newValue }
+        // swiftlint:disable no_breadcrumb_data_setter
+        set { wrapped.data = newValue } // `wrapped` is a Mechanism, not a Breadcrumb
+        // swiftlint:enable no_breadcrumb_data_setter
     }
 
     @objc public var handled: NSNumber? {