+++ Tests/SentryTests/Integrations/WatchdogTerminations/Processors/SentryWatchdogTerminationBreadcrumbProcessorTests.swift
@@ -13,6 +13,8 @@ class SentryWatchdogTerminationBreadcrumbProcessorTests: XCTestCase {
         let currentDate = TestCurrentDateProvider()
         let maxBreadcrumbs = 10
 
+        // swiftlint:disable no_breadcrumb_data_setter
+        @available(*, deprecated, message: "Testing deprecated Breadcrumb.data setter")
         init() throws {
             breadcrumb = TestData.crumb
             breadcrumb.data = nil
@@ -27,6 +29,7 @@ class SentryWatchdogTerminationBreadcrumbProcessorTests: XCTestCase {
                 dispatchQueueWrapper: TestSentryDispatchQueueWrapper()
             )
         }
+        // swiftlint:enable no_breadcrumb_data_setter
 
         func getSut() -> SentryWatchdogTerminationBreadcrumbProcessor {
             return getSut(fileManager: self.fileManager)
@@ -42,6 +45,7 @@ class SentryWatchdogTerminationBreadcrumbProcessorTests: XCTestCase {
     private var fixture: Fixture!
     private var sut: SentryWatchdogTerminationBreadcrumbProcessor!
 
+    @available(*, deprecated, message: "Testing deprecated Breadcrumb.data setter")
     override func setUpWithError() throws {
         try super.setUpWithError()